export column diffing
