+++ push.go
@@ -7,7 +7,9 @@ import (
 	"strings"
 	"time"
 
+	"github.com/clearblade/cblib/colutil"
 	"github.com/clearblade/cblib/internal/types"
+	"github.com/clearblade/cblib/listutil"
 	"github.com/clearblade/cblib/models"
 	"github.com/clearblade/cblib/models/bucketSetFiles"
 	libPkg "github.com/clearblade/cblib/models/libraries"
@@ -158,7 +160,7 @@ func pushUserSchema(systemInfo *types.System_meta, client *cb.DevClient) error {
 		return fmt.Errorf("Error in schema definition. Pls check the format of schema...\n")
 	}
 
-	diff := getDiffForColumnsWithDynamicListOfDefaultColumns(localSchema, userColumns)
+	diff := colutil.GetDiffForColumnsWithDynamicListOfDefaultColumns(localSchema, userColumns)
 	for i := 0; i < len(diff.Removed); i++ {
 		if err := client.DeleteUserColumn(systemInfo.Key, diff.Removed[i].(map[string]interface{})["ColumnName"].(string)); err != nil {
 			return fmt.Errorf("User schema could not be updated. Deletion of column(s) failed: %s", err)
@@ -170,25 +172,6 @@ func pushUserSchema(systemInfo *types.System_meta, client *cb.DevClient) error {
 		}
 	}
 	return nil
-}
-
-func getDiffForColumnsWithStaticListOfDefaultColumns(localSchemaInterfaces, backendSchemaInterfaces []interface{}, defaultColumns []string) *UnsafeDiff {
-	return compareListsAndFilter(localSchemaInterfaces, backendSchemaInterfaces, columnExists, func(a interface{}) bool {
-		return !isDefaultColumn(defaultColumns, a.(map[string]interface{})["ColumnName"].(string))
-	})
-}
-
-func getDiffForColumnsWithDynamicListOfDefaultColumns(localSchemaInterfaces, backendSchemaInterfaces []interface{}) *UnsafeDiff {
-	return compareListsAndFilter(localSchemaInterfaces, backendSchemaInterfaces, columnExists, func(a interface{}) bool {
-		return a.(map[string]interface{})["UserDefined"].(bool)
-	})
-}
-
-func columnExists(colA interface{}, colB interface{}) bool {
-	if colA.(map[string]interface{})["ColumnName"].(string) == colB.(map[string]interface{})["ColumnName"].(string) && colA.(map[string]interface{})["ColumnType"].(string) == colB.(map[string]interface{})["ColumnType"].(string) {
-		return true
-	}
-	return false
 }
 
 func pushEdgesSchema(systemInfo *types.System_meta, client *cb.DevClient) error {
@@ -207,7 +190,7 @@ func pushEdgesSchema(systemInfo *types.System_meta, client *cb.DevClient) error 
 		return fmt.Errorf("Error in schema definition. Please verify the format of the schema.json. Value is: %+v - %+v\n", edgeschema["columns"], ok)
 	}
 
-	diff := getDiffForColumnsWithDynamicListOfDefaultColumns(typedLocalSchema, allEdgeColumns)
+	diff := colutil.GetDiffForColumnsWithDynamicListOfDefaultColumns(typedLocalSchema, allEdgeColumns)
 	for i := 0; i < len(diff.Removed); i++ {
 		if err := client.DeleteEdgeColumn(systemInfo.Key, diff.Removed[i].(map[string]interface{})["ColumnName"].(string)); err != nil {
 			return fmt.Errorf("Unable to delete column '%s': %s", diff.Removed[i].(map[string]interface{})["ColumnName"].(string), err.Error())
@@ -238,7 +221,7 @@ func pushDevicesSchema(systemInfo *types.System_meta, client *cb.DevClient) erro
 		return fmt.Errorf("Error in schema definition. Please verify the format of the schema.json\n")
 	}
 
-	diff := getDiffForColumnsWithDynamicListOfDefaultColumns(localSchema, allDeviceColumns)
+	diff := colutil.GetDiffForColumnsWithDynamicListOfDefaultColumns(localSchema, allDeviceColumns)
 	for i := 0; i < len(diff.Removed); i++ {
 		if err := client.DeleteDeviceColumn(systemInfo.Key, diff.Removed[i].(map[string]interface{})["ColumnName"].(string)); err != nil {
 			return fmt.Errorf("Unable to delete column '%s': %s", diff.Removed[i].(map[string]interface{})["ColumnName"].(string), err.Error())
@@ -1360,7 +1343,7 @@ func pushCollectionSchema(systemInfo *types.System_meta, collection map[string]i
 		return fmt.Errorf("Error in schema definition. Please verify the format of the schema.json\n")
 	}
 
-	diff := getDiffForColumnsWithStaticListOfDefaultColumns(localSchema, backendSchema, DefaultCollectionColumns)
+	diff := colutil.GetDiffForColumnsWithStaticListOfDefaultColumns(localSchema, backendSchema, DefaultCollectionColumns)
 	for i := 0; i < len(diff.Removed); i++ {
 		if err := cli.DeleteColumn(collID, diff.Removed[i].(map[string]interface{})["ColumnName"].(string)); err != nil {
 			return fmt.Errorf("Unable to delete column '%s': %s", diff.Removed[i].(map[string]interface{})["ColumnName"].(string), err.Error())
@@ -1482,8 +1465,8 @@ func updateDeployment(systemInfo *types.System_meta, cli *cb.DevClient, name str
 }
 
 func diffDeployments(localDep map[string]interface{}, backendDep map[string]interface{}) map[string]interface{} {
-	assetDiff := compareLists(localDep["assets"].([]interface{}), backendDep["assets"].([]interface{}), isAssetMatch)
-	edgeDiff := compareLists(localDep["edges"].([]interface{}), backendDep["edges"].([]interface{}), isEdgeMatch)
+	assetDiff := listutil.CompareLists(localDep["assets"].([]interface{}), backendDep["assets"].([]interface{}), isAssetMatch)
+	edgeDiff := listutil.CompareLists(localDep["edges"].([]interface{}), backendDep["edges"].([]interface{}), isEdgeMatch)
 	return map[string]interface{}{
 		"assets": map[string]interface{}{
 			"add":    assetDiff.Added,
