+++ utils.go
@@ -14,8 +14,6 @@ import (
 	"reflect"
 	"strings"
 
-	d "github.com/clearblade/cblib/diff"
-
 	cb "github.com/clearblade/Go-SDK"
 	"github.com/clearblade/cblib/internal/types"
 )
@@ -348,36 +346,6 @@ func removeBlacklistedPortalKeys(portal map[string]interface{}) map[string]inter
 type ListDiff struct {
 	add    []interface{}
 	remove []interface{}
-}
-
-func isDefaultColumn(defaultColumns []string, colName string) bool {
-	for i := 0; i < len(defaultColumns); i++ {
-		if defaultColumns[i] == colName {
-			return true
-		}
-	}
-	return false
-}
-
-// FilterSlice returns the items of the slice `s` for which `predicate` returns true.
-func FilterSlice(s []interface{}, predicate func(interface{}) bool) []interface{} {
-	filtered := make([]interface{}, 0, len(s))
-
-	for _, item := range s {
-		if predicate(item) {
-			filtered = append(filtered, item)
-		}
-	}
-
-	return filtered
-}
-
-func compareListsAndFilter(after []interface{}, before []interface{}, compare func(interface{}, interface{}) bool, filter func(interface{}) bool) *d.UnsafeDiff {
-	diff := d.UnsafeDiff{after, before, nil, nil, compare}
-	d.Diff(&diff)
-	diff.Added = FilterSlice(diff.Added, filter)
-	diff.Removed = FilterSlice(diff.Removed, filter)
-	return &diff
 }
 
 func convertStringSliceToInterfaceSlice(strs []string) []interface{} {
